lib/titiler-pgstac-api/runtime/requirements.txt
titiler.pgstac==0.3.3
boto3>=1.26.139
psycopg[binary, pool]